finish path_construct

+++ Ant.py
@@ -5,7 +5,7 @@ from City import *
 class Ant(object):
 
     def __init__(self):
-        self.path = np.zeros(1)
+        self.path = np.zeros(1, dtype=int)
         self.dis = 999999999
 
     def set_path(self, key, val):
